compile honey-badger.exe
