+++ make-honey-badger.lst
@@ -34,3 +34,5 @@ C:\Program Files\IronPython 2.7\Lib\functools.py
 C:\Program Files\IronPython 2.7\Lib\shutil.py
 C:\Program Files\IronPython 2.7\Lib\fnmatch.py
 C:\Program Files\IronPython 2.7\Lib\traceback.py
+C:\Program Files\IronPython 2.7\Lib\csv.py
+C:\Program Files\IronPython 2.7\Lib\importlib\__init__.py
